Final Modified CMS

+++ FRONTEND/src/app/components/manageusers/manageusers.component.ts
@@ -4,6 +4,7 @@ import { Tpo } from '../../models/tpo';
 import { AuthService } from '../../services/auth.service';
 import { FlashMessagesService } from 'angular2-flash-messages';
 import { Router } from '@angular/router';
+import {HeaderComponent} from '../header/header.component';
 
 
 @Component({
@@ -14,7 +15,7 @@ import { Router } from '@angular/router';
 export class ManageusersComponent implements OnInit {
   users: any = [];
   type:String;
-  usertype:String;
+  usertype:String="USERS";
   constructor(private authService:AuthService,
     private flashmessage:FlashMessagesService,
     private router:Router) {
@@ -29,11 +30,12 @@ export class ManageusersComponent implements OnInit {
     this.authService.userTypesObservable$.subscribe(
       (type: String) => {
          this.getUsers(type);
+         this.type=type;
       }
     )
   }
-  hodcomp(type:String){
-    if(type=="hod"){
+  hodcomp(){
+    if(this.type=="hod"){
       return true;
     }
     else{
@@ -41,10 +43,14 @@ export class ManageusersComponent implements OnInit {
     }
   }
   getUsers(type: String){
-    this.hodcomp(type);
+    this.hodcomp();
     this.usertype=type+"s";
     this.users = [];
-    type === 'hod' ? this.managehod() : ( type === 'tpo' ? this.managetpo() : '') 
+    type === 'hod' ? this.managehod() : ( type === 'tpo' ? this.managetpo() :(type === 'undefined' ? this.emptyuser() : '')) 
+  }
+  emptyuser(){
+    this.usertype="USERS";
+    this.users=[];
   }
   managehod(){
     console.log("come to this area HOD");
@@ -71,22 +77,29 @@ export class ManageusersComponent implements OnInit {
       });
   }
   delete(hod){
-    console.log(hod);
-    this.authService.deleteUser(hod.userid).subscribe(data=>{
-      if(data.success){
-        this.flashmessage.show("student record deleted",{cssClass:'alert-success text-center',timeOut:2000});
-        this.users.splice(this.users.indexOf(hod),1)
-      }
-      else
-      {
-        console.log(data);
-        this.flashmessage.show("Something went wrong.",{cssClass:'alert-danger text-center',timeOut:2000});
-      }
-    })
+    //console.log(hod);
+    var retVal = confirm("Are you sure to Delete?");
+        if( retVal == true ){
+          this.authService.deleteUser(hod.userid).subscribe(data=>{
+            if(data.success){
+              this.flashmessage.show("student record deleted",{cssClass:'alert-success text-center',timeOut:2000});
+              this.users.splice(this.users.indexOf(hod),1)
+            }
+            else
+            {
+              console.log(data);
+              this.flashmessage.show("Something went wrong.",{cssClass:'alert-danger text-center',timeOut:2000});
+            }
+          })
+        }
+        else{
+        console.log ("Admin doesn't want to Delete!");
+        //this.authService.toggleForm=!this.authService.toggleForm;
+        this.router.navigate(['/manageusers']);  
+        }
   }
   update(hod){
       this.authService.selectedUser=hod;
       this.authService.toggleForm=!this.authService.toggleForm;
-    }
-   
+  }
 }
